+++ src/components/DataGrid/index.ts
@@ -1,1 +1,3 @@
 export * from './StyledDataGrid'
+export * from './StyledGridActionsCellItem'
+export * from './x-data-grid-types'